long type should be int values; implement more date/time and misc other functions
git-svn-id: https://svn.code.sf.net/p/jackcess/code/jackcess/branches/exprs@1120 f203690c-595d-4dc9-a70b-905162fa7fd2

+++ src/main/java/com/healthmarketscience/jackcess/impl/expr/ExpressionTokenizer.java
@@ -394,7 +394,7 @@ class ExpressionTokenizer
         // what number type to use here?
         Object num = (isFp ? 
                       (Number)Double.valueOf(numStr) : 
-                      (Number)Long.valueOf(numStr));
+                      (Number)Integer.valueOf(numStr));
         foundNum = true;
         return new Token(TokenType.LITERAL, num, numStr, 
                          (isFp ? Value.Type.DOUBLE : Value.Type.LONG));
